Move rewards computation into inference submodule
